fix UI: completed split of Countdown Time component and it does not appear in select ticket page
