+++ frontend/src/app/(event)/event/[id]/booking/question-form/page.tsx
@@ -1,15 +1,19 @@
 'use client'
 
+/* Package System */
 import React from 'react';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import 'bootstrap-icons/font/bootstrap-icons.css';
 import 'tailwindcss/tailwind.css';
+import { useEffect, useState } from 'react';
+
+/* Package Application */
 import '@/styles/admin/pages/Dashboard.css';
 import '@/styles/admin/pages/BookingQuestionForm.css'
-import { useEffect, useState } from 'react';
 import QuestionList from './components/questionList';
 import TicketInformation from './components/ticketInfo';
 import Navigation from '../components/navigation';
+import CountdownTimer from '../components/countdownTimer';
 
 
 export default function QuestionForm() {
@@ -41,12 +45,16 @@ export default function QuestionForm() {
     return (
         <div className="mt-5 mb-5">
             <Navigation title="Bảng câu hỏi" />
+            
+            <div className="fixed top-10 right-10 mt-4">
+                <CountdownTimer />
+            </div>
+
             <div className="px-32 py-0">
                 <div className="row align-items-start mt-4">
                     <QuestionList onValidationChange={setIsFormValid} />
                     <TicketInformation event={event} totalTickets={totalTickets} totalAmount={totalAmount} isFormValid={isFormValid} ticketTypeName={ticketTypeName} ticketPrice={ticketPrice} />
                 </div>
-
             </div>
         </div>
     )
